Updated components: • Made Strikes a simpler fn component • Separated components from non-component JS • Use React.memo for fn components
