+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDom from 'react-dom';
-import App from './App';
+import Game from './components/Game';
 
 ReactDom.render(
-  <App />,
+  <Game />,
   document.getElementById('root'),
 );
